# rezara-app/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Booking Confirmation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: getReservations
      tags: [reservations]
      summary: List reservations for the current business
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_payment, confirmed, cancelled, completed, expired]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a new reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationBody"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{reservationId}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation by ID (public, for customer payment page)
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation with business info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationWithBusiness"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateReservation
      tags: [reservations]
      summary: Update a reservation
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationBody"
      responses:
        "200":
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: cancelReservation
      tags: [reservations]
      summary: Cancel a reservation
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{reservationId}/complete:
    post:
      operationId: completeReservation
      tags: [reservations]
      summary: Mark reservation as completed
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{reservationId}/status:
    patch:
      operationId: changeReservationStatus
      tags: [reservations]
      summary: Change the status of a reservation
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeReservationStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Business:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        logo:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - createdAt
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        linkId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
          nullable: true
        date:
          type: string
        time:
          type: string
        guests:
          type: integer
        depositAmount:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending_payment, confirmed, cancelled, completed, expired]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - linkId
        - customerName
        - date
        - time
        - guests
        - depositAmount
        - status
        - createdAt
        - updatedAt
    ReservationWithBusiness:
      allOf:
        - $ref: "#/components/schemas/Reservation"
        - type: object
          properties:
            business:
              $ref: "#/components/schemas/Business"
          required:
            - business
    CreateReservationBody:
      type: object
      properties:
        linkId:
          type: string
          description: Optional client-generated link ID (10-char alphanumeric). Server generates one if omitted.
        customerName:
          type: string
        customerPhone:
          type: string
          nullable: true
        date:
          type: string
        time:
          type: string
        guests:
          type: integer
        depositAmount:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - customerName
        - date
        - time
        - guests
        - depositAmount
    UpdateReservationBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
          nullable: true
        date:
          type: string
        time:
          type: string
        guests:
          type: integer
        depositAmount:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending_payment, confirmed, cancelled, completed, expired]
...
    ChangeReservationStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending_payment, confirmed, cancelled, completed, expired]
      required:
        - status
...